useReducer() recap | React hooks 2021
