+++ src/ToDo/constants.js
@@ -1,0 +1,3 @@
+export const SET_JOB = "set_job";
+export const ADD_JOB = "add_job";
+export const DELETE_JOB = "delete_job";
